Do not enforce response token if it' the RPC itself failed
Summary:
We shouldn't be enforcing on tokens here because tokens are not there. The idea is that in this case we never reached the destination anyway, so it's pointless to be verifying responder tokens.

Skipping enforcement here allows us to not override RPC problems with token mismatches, which can shortcircuit any error handling.

Test Plan: Upload to merlin. The easiest way to check this is to stop a follower mysqld and start an election somewhere else. The election protocol will attempt to request votes from this stopped instance, so check that it doesn't print a mismatch.

Reviewed By: alanliang

Differential Revision: D39835931

fbshipit-source-id: 70b2a966eadc2a3a3c431a3039337f8eb555138f