Remove the erase for next leader regions
Summary:
Just based on counting votes, we cannot determine that we can skip a region. This was the cause for S296607, where on term x the leader was in FTW, term x+1 votes point to PNB but the election wasn't won, and term x+2 only tried to get majority from PNB.

Even in the case where election was clearly won in PNB, we can't skip FTW, because if a majority of PNB is lagging and the PNB leader wasn't able to commit no-op for any reason (crashes, failed mysql settings), then PNB can certify a next leader to be ahead incorrectly.

The original intent for voter history was the the regions to get majority from can only grow with each term we look at until we end up with the pessimistic quorum.

Test Plan:
Added a flexible vote counter test to check that we can't win an election without majority from the middle term.

Verified that these tests fail without this diff.

Reviewed By: iRitwik

Differential Revision: D39836378

fbshipit-source-id: 0eb19d3db20a4196fb1dae6170e83473a9f734fe