Move definition of ElectionMode to protobuf and add MOCK_ELECTION
Summary:
We need this in protobuf because we need to be able to differentiate the election mode within VoteRequest. Currently, we have `VoteRequest.is_pre_election` but adding another `is_mock_election` is not clean enough and can be confusing. It will also make it harder in the future if we add more modes.

Note, this diff does not change any logic. We will utilize MOCK_ELECTION in later diffs.

Test Plan: Existing tests.

Reviewed By: yichenshen

Differential Revision: D40309310

fbshipit-source-id: c79a20d7f74348f131e15683f5c297312d4e6aa4