Converting spin locks in RaftConsensus and PeerMessageQueue to std::mutex
Summary:
In mysqlslap benchmarks spin locks were taking about 1% CPU on the
commit path. When converted to regular mutex this overhead virtually disappeared.

The main reason I think is because we hold these locks during append and
we don't append to log asynchronously (like kudu). So the critical section for
these locks is pretty large and spin locks are not good in that case.

Test Plan: buck tests

Reviewed By: alanliang, anirbanr-fb

Differential Revision: D39756484

fbshipit-source-id: fc452cc6e3dee692013d3074543be26ece59c19f