Track when we have all voters in for Flexible vote counter
Summary:
Track how many members we have in each quorum id/region. This is taken from the committed config.

Now if voter distribution says that a quorum's distribution is X, but we have Y < X members, once we get Y votes we can make a conclusion on whether quorum can be satistifed based on if it is satistifed based off the Y votes.

This avoids cases where elections waits forever for X votes, when there will never be X votes.

We fix the very base counter logic by comparing vote counts with total members. If they are equal, we make quorum_satisfaction_possible take the value of quorum_satisfied because no more votes can come in to change the result.

> I also altered the logging a bit. Now it DFATALs on mismatch between satistifed and possible when all votes are in. In prod DFATAL is treated as ERROR.

Test Plan: Added a Flexible Voter test. Verified that it fails before and suceeds now.

Reviewed By: abhinav04sharma

Differential Revision: D40437411

fbshipit-source-id: 5f840e21aa31db54214bf9ff3fc0481b7b01df17