New var to specify min threads in the raft thread pool
Summary:
We already have one for max threads, adding a new one for min threads.
The 1st min threads in the pool are marked as permanent and cannot be reaped.
They use pthread_cond_wait instead of pthread_cond_timed_wait which saves 0.5%
CPU in mysqlslap benchmarks.

Setting the default value of min threads as 8 which is half of the 16 max
threads that we usually spawn today with a 1 hour idle time.

Test Plan: buck/cogwheel

Reviewed By: anirbanr-fb

Differential Revision: D40578402

fbshipit-source-id: a8f4bafee9a5f529fcfef34b77216299ad00f2f8