Allow voting in mock elections without update lock
Summary:
We have to code to reject votes if we don't get an update lock. For mock elections this prevents the current leader from voting for the candidate.

So it has to be relaxed for mock elections too.

Test Plan: This can be tested by uploading to merlin and running a couple of LMPs. Check that in each case the existing leader votes for the candidate.

Reviewed By: alanliang

Differential Revision: D40745523

fbshipit-source-id: dad147c5de0695bc20ef4cbe4c624f7c74a59173