Propagate error response when elections are disabled.
Summary:
yichenshen pointed out that when elections are disabled, mock elections RPCs time out after 30 seconds instead of erroring out immediately. The main cause is that we were returning Status::OK() when elections are disabled and because mock elections wait for a response, a response is never returned and it times out after 30 seconds.

Upon fixing this and writing a test, I also found the issue that RunLeaderElection responses were not being propagated properly back to the caller since we were returning the ConsensusResponsePB instead of the RunLeaderElectionResponsePB. This causes us to always return a runtime error with unknown error codes.

This led to the discovery of another problem: LOG_WITH_PREFIX_UNLOCKED called without lock.

Test Plan:
- Added tests.
- Tested in merlin by disabling rpl_raft_allow_start_election sys var on candidate, changed leader to candidate and observed proper failure on leader.

Reviewed By: yichenshen

Differential Revision: D41315713

fbshipit-source-id: a9288f4dd210d2fa397eb3bddb1b1046b8bdfb0f