Add kudu nalespace to base::internal namespace offering SpinLock capabilities: prevent symbols duplication (if other dependencies are vendoring gutil)
Summary:
kuduraft vendors gutil library. This implies that if the library is used in conjunction of other libraries vendoring gutil and/or importing gutil we would have the same symbol defined multiple times.
In such conditions, depending on compiler/linker flags we might end up with a build failure or with one of the vendored gutil actually present on the final binary, which might nullify possible changes on the vendored gutil version.

This change aims to create a sub-namespace for the vendored gutil such that symbol duplication would not be possible.

Reviewed By: yichenshen

Differential Revision: D41253791

fbshipit-source-id: 1aa8e17ce5e8fb3e6ff39cb4353f2a22667c5773