Allow non-empty root directory while initing kudu raft
Summary:
FsManager::CreateFileSystemRoots expects the root directory to be empty.
While this works for mysql raft, on zippydb it would be nice to start with
a root directory that's currently used to keep other persistent state.
Provide an option to relax the "empty root" requirement so, that FsManager
can continue to setup the fs layout and directories it needs in the root
directory provided to it. ZippyDB will need to ensure there isn't any name
collision.

Test Plan:
- compile
- mysql raft tests pass; The change is also a no-op on mysql
- zippydb tests in a stacked diff

Reviewed By: dyniusz, yichenshen

Differential Revision: D42060272

fbshipit-source-id: 2e07f461518943149007c8d685615f233c57db77