Track number of consecutive rpc failures per peer and define a failure threshold for a peer to be considered unhealthy
Summary:
There are many places where we need to determine whether a peer is healthy or not. Use cases include T136385187 and  T143022481. This diff adds tracking of the number of consecutive rpc failures per peer and based on a configurable threshold, we can deem a peer healthy or not.

I also considered the existing attributes to infer availability:
- last_communication_time
- last_successful_exchange
- last_exchange_status
- last_received (opid)

However, anything time related or last status related didn't provide a clear picture of availability. In some ways, these heuristics could be too sensitive. Counting the number of consecutive failures provides the right amount of slack to infer the health of a peer.

Test Plan: Tested in subsequent diffs

Reviewed By: yichenshen

Differential Revision: D42817847

fbshipit-source-id: 1a3aa78342e2fd9208d4d1d5fb246e3c98f80272