Attempt to transfer leadership when leader cannot make progress with majority
Summary:
When a leader fails committing to a quorum of peers, we want to be able to recover from this case and transfer leadership to another peer. Today, when we fail to have a commit quorum, the leader will not resolve this on its own. This requires human intervention to transfer the leadership.

In this diff, we create a process called "Check Quorum", that runs at an interval, checking to see if the current leader can commit to its quorum. If it cannot, we attempt to transfer the leadership to another peer. This transfer of leadership will be an LMP and will utilize Mock Elections.

Rollout plan:

Initial flags:
  - `rpl_raft_check_quorum = false`
  - `rpl_raft_check_quorum_failure_callback = false`

1. `rpl_raft_check_quorum` will be progressively set to true
2. Check metrics and make sure the cases we are failing check quorum are sane.
3. `rpl_raft_check_quorum_failure_callback` will be set to true.

Test Plan:
- Added unit tests
- Tested in merlin
  - Tested by breaking quorum using `rpl_raft_reject_append_entries` and observed that check quorum failure callback is called leadership is transferred
  - Tested breaking a single peer and ensured that no failure callback is called since quorum is intact.
  - Tested enabling check quorum on a follower and ensured there was no side effects.
  - Tested disabling check quorum after it was turned on
  - Tested enabling check quorum but disabling check quorum failure callback
  - Tested changing `rpl_raft_check_quorum_interval_ms` and observed that it was effective
  - Tested changing all other vars

Reviewed By: yichenshen

Differential Revision: D42252951

fbshipit-source-id: 04718d2a63eb79e64bc438c9d7d5d58f5317e973