Add metric for available commit peers
Summary:
In previous diff, we created a status variable to track number of available commit peers (healthy peers that are part of the commit quorum).

In this diff, we emit this variable as a gauge metric. We will piggy back on the periodic Check Quorum thread to update the gauge metric.

Test Plan: Deployed to merlin and verified that new column is emitted: https://fburl.com/scuba/mysql_raft_metrics/2p6roj4w

Reviewed By: yichenshen

Differential Revision: D42972397

fbshipit-source-id: fde1dc15a6b69541336909d4d837014f05bd827a