Status variable to return number of available commit peers
Summary:
Add a status variable that returns the number of healthy peers that are part of the commit path. A peer is unhealthy when it exceeds a configured amount of consecutive rpc errors (any non-OK controller status). See D42817847 (https://github.com/facebook/kuduraft/commit/48100a5cb719e8eb2408d59f8cc8c8e4bf2ac5d5) for implementation of that.

The goal here is to understand how far we are from being a single point of failure.

In the next diff, I will emit this status var as a metric so that we can dive deeper with scuba.

Test Plan:
- Added assertions to existing tests
- Tested in Merlin. Ran `SHOW RAFT STATUS` and ensured that status var populated correctly for leader and follower.

Reviewed By: yichenshen

Differential Revision: D42950022

fbshipit-source-id: 576eda86147aa803f9c57539192b63818ba595a5