Change check quorum interval setting from milliseconds to number of heartbeats
Summary: This sys var controls the interval at which we run check quorum. Instead of defining the interval in terms of milliseconds, it's better to define it in terms of heartbeats because the status variables that check quorum relies on are updated on the order of heartbeats.

Test Plan: deployed to merlin and tried setting the sysvar

Reviewed By: yichenshen

Differential Revision: D43683157

fbshipit-source-id: dfe5d62966f72e3daad9488abef00c6361a95952