New flag to track if LDCB or NORCB was called at least once
Summary:
In the earlier code, if there was an error in UpdateReplica after leader uuid is
populated but before LDCB is called (e.g. a compression error, checksum error
etc.) we ended up in a state where LDCB would never be called and so the
instance would not be setup properly. To fix this introduced a new flag that
tracks if LDCB has been called at least once.

Test Plan:
Tested on stage -1

LDCB was called after compression error on a new replica:
{P646400668}

Restarts also trigger LDCB:
{P647544557}

Reviewed By: yichenshen

Differential Revision: D43882986

fbshipit-source-id: 1c84db58e2ad0e160ad5b4b752d245a95da1c4bf