Sending compression dictionary in 1st request by default
Summary:
Setting `should_send_compression_dict` to true by default so that we send
compression dictionary to new peers in the 1st request (even if the dictionary
is empty). This flag is reset on 1st successful response. This way we don't have
to wait for a request to fail due to compression before sending dict to a newly
added node. For compression codecs that don't support dictionaries (e.g. LZ4)
this will be a no-op.

Also, copying over the dict in downstream proxy request.

Test Plan:
Tested on devserver, on leader:
{P651975573}

On follower:
{P651978015}

Reviewed By: li-chi

Differential Revision: D44063891

fbshipit-source-id: 99f6b3fa8117372d8a6b781add3ed5ed72f0c856