Remove erroring when quorum health is retrieved on a follower
Summary:
When SHOW RAFT STATUS is requested on a follower, we currently log:
```
E0317 13:15:12.264904 33226 MysqlRaft.cpp:4954] Failed to get quorum health: Not found: Quorum health can only be retrieved on a leader
```

This can be quite noisy in the logs as db status constantly is querying SHOW RAFT STATUS on followers. Instead of failing, we should return success. This is fine because the resultant response will be empty anyway.

Test Plan:
Existing tests.

Tested in merlin and initiated SHOW RAFT STATUS on a follower and observed no failure.

Reviewed By: li-chi

Differential Revision: D44180392

fbshipit-source-id: 93b1b5bd6814827cef9d8d7a770265e6c792937e