Handling bootstrap info better
Summary:
In D21133184 we made a change to how bootstrap info is handled to fix a
double free problem. This changed GetRecoveryInfo() such that it can only be
called once (because we use std::move in it). This creates confusion in
ownership and potential to write buggy code in the future.

Three main changes were made:
1. Orphaned replicate vector now contains shared_ptr of ReplicateMsgs
2. Bootstrap into itself is maintained as a shared_ptr in kudu:Log
3. Log can be considered the owner of bootstrap object. Others use
Log::GetRecoveryInfo() to get the bootstrap obj.

Test Plan: buck/cogwheel

Reviewed By: anirbanr-fb

Differential Revision: D45127382

fbshipit-source-id: 9c820a8c060f8cd41737ab1193c734d8b9adbc0a