Simply potential leader append
Summary:
This is a logical refactor.

We loop through leader regions and add the candidate to a set if we find a majority in any leader region.

We look at 2 conditions and do the same things.

Combine the 2 conditions and return the function after we add the candidate (adding it more times doesn't do anything).

Test Plan: sandcastle tests

Reviewed By: li-chi

Differential Revision: D46090800

fbshipit-source-id: 44157aa92d388dbd149c7409a603d1c0b4ac06cf