Prune history properly
Summary:
I discovered this when testing the previous diff.

We're basically not pruning history at all. The protobuf map is actually documented as an unordered map: https://protobuf.dev/reference/cpp/api-docs/google.protobuf.map/

The actual iteration order the current implementation produces is actually in reverse order of insertion.

As such, the current pruning logic will always find the very last vote and just bail out.

When it hits the max count, then it'll essentially trim the first item and set that as the pruned term. I.e. it'll always prune everything once it hits a hundred. Voter history would be overly pessimistic in such cases. We've not seen Instances almost never survive a hundred votes in prod, so quite unlikely we hit that path.

Test Plan: I added some logging on how we're pruning the history and how we're reading them on the candidate side. Uploaded this to merlin and ran a few promotions to ensure everything's sane.

Reviewed By: iRitwik

Differential Revision: D46204090

fbshipit-source-id: 1ffaa151be946392661869e753e9d7306980870e