Consider unpopulated voter history to be pruned
Summary:
Voter history on voters can be pruned by the voters themselves. This happens when it receives a LKL on a higher term, or if it hit more than 100 terms: https://fburl.com/code/ybd8l8vv

Now if history is pruned, voter history can wrongly conclude that the vote was not given (because there's no history). This is why we always populate a last pruned term when sending voter history: https://fburl.com/code/tb55unh9

When looking at voter history, pruned terms are treated conservatively, and will count as "yes" votes in the quorum discovery phase. The idea is that if we don't know if the voter has voted or not, we assume it has so that we conservatively add it's quorum for intersection.

There's an edge case here when the rpc itself has failed. In this case, we get neither pruned term nor voter history, and in this case we could potentially miss considering some quorums. This diff changes it such that we check if the pruned term is calculated, and if not, bump it up all the way to the election term to signify that we didn't get any sort of voter history information out of the voter at all.

Today, because we only look at voter histories of quorums we already need a intersection for, it's very unlikely to cause problems, since the final quorum computation would already have failed. There is an edge case with split votes + network cuts that can be a problem today, so this should be fixed.

Test Plan: Added some logging to this codepath. Uploaded to merlin and brought down a member. Ensure that the pruned term is overriden correctly.

Reviewed By: iRitwik

Differential Revision: D46193459

fbshipit-source-id: 8d4b6dcc432210f65b1e9017cb040d01b71b2d9d