Remove the socket buffer settings
Summary:
I added this during one of the old sevs where Raft would suddenly lag on replicas. This never worked to begin with, because setting the buffer at this point in the connection actually made lag worse (Domas told be it's due to some buffer clamp setting, meaning we actually end up with a tiny buffer if we try to set this). The buffer is better off being adjusted by the kernel dynamically.

Remove this. The original lag issue was due to packet drops, and is solved by doing ICP tagging.

Test Plan: contbuild

Reviewed By: li-chi

Differential Revision: D46172433

fbshipit-source-id: 7d06daf660d9a92890861afff0edd4afd678e9af