Removed explored leader regions
Summary:
We used to remove leader regions under some conditions. The explored regions set was meant to account for that so we persist the ones that have already been scanned.

However, since D39836378, we found that removing leader regions is unsafe, and as such, the next_leader_region list can now only grow. So we don't need the explored regions set anymore.

Test Plan: FlexibleVoterCounterTests

Reviewed By: alanliang

Differential Revision: D46091952

fbshipit-source-id: cd61022d51865aff243b25cb1cf26b6bc20d73f4