Local tracked peer should have different default state
Summary:
Tracking the local peer (aka the Leader) is a special case because we
do not update the tracked peer's state during replication. However, we do
rely on the the tracked local peer's state, such as consecutive_failures, for
CheckQuorum. In CheckQuorum, we always assume the local peer is healthy and so
consecutive_failures counters should always be 0.

Hence, there should be a way to initialize local tracked peers with its own defaults.

Test Plan: Existing tests

Reviewed By: yichenshen

Differential Revision: D46460524

fbshipit-source-id: b6b95be52344b7922c9530a98ecb07ca5f472018