Evaluate candidate quorum before adding to potential leader quorum
Summary:
We consider candidate quorum when running elections. However, voter history currently does not consider candidate quourm votes when looking for potential leader quorums.

This means that it will include a potential leader quorum even when it's certain that the candidate in that quorum would not have won. We add a check to the path where we consider a potential leader to look for candidate quorum votes as well.

The logic here is conservative, and it treats all unknown votes as yes votes when determining if we need to ask for a majority, which means in the case where some candidates don't respond, or have trimmed their histories, we would simply consider them as agreeing that we should ask that quorum for a majority even if might turn out that we might not need to. In order words, in the case where vote history is lacking, we err on the side of safety.

Test Plan: FlexibleVoteCounterTests have various scenarios that should still pass. Added another one for candidate regions.

Reviewed By: anirbanr-fb

Differential Revision: D46173790

fbshipit-source-id: 05b8cd757d24b68168555ca7e2c3c3419b1e21aa