Make voter history more optimistic about being able to eventually achieve majority
Summary:
Voter history has 2 stages. First we scan historical votes to find a list of quorums to intersect with. Then we look to see if we have votes currently from these quorums.

Today, we also decide if it's still possible to win the election based on the remaining votes left for these quorums. For example, if at some point we determine we should intersect with A, B and C, and we didn't get enough votes in C, we'll check if there are still enough votes left to get enough votes in C. If not, we call the election lost.

In the voter history scan though, when we don't have enough information, we act conservatively by adding the quorum. For example, if we tallied that for term X, we have 1/3 votes for C 1/3 against and 1/3 that hasn't come in, we'll add C into the potentially leader quorums since we don't know if the last vote will be for C or not, so we just take a majority from C for the vote tally on the current election to be safe.

The issue however is that if we can't get enough votes from C itself on our election, we'll immediately declare the election lost. But if the last vote for term X for C comes in later and it says that it was against C, then in reality we don't need to intersect with C at all.

As such, in this diff we keep track if we added a potential leader quorum using unreceived record. Since this unreceived record could later turn out to be a no vote, we shouldn't call the election early if we used them to add potential leader regions. It is also guaranteed that when all votes are in, we will have no unreceived records, so we'll always be able to make a final call on the election at that point.

> One thing to take note here is the difference between an unreceived record and a undecidable record. An unreceived record is a vote count that we have not yet received, since the instance providing us with the record has not yet responded. An undecidable record is a record we know we cannot get, like from a RPC error, or if it was pruned on the providing instance. Both count as yes votes in determining potential leaders in order to be conservative, but only unreceived records can later
> change. And as such, we only hold off calling the election if we used unreceived records, not if we used undecidable records.

Test Plan:
This is such an obscure case that it'll be really hard to repro with an actual running ring.

Instead, add some tests to the voter counter test.

Verified that before this change the test will fail. It's obvious that in the test case the first `canAchieveMajority` should be true, since registering the last vote actually allows it to achieve the majority.

Reviewed By: alanliang

Differential Revision: D46747766

fbshipit-source-id: 5d82df740a1197d2a8a80aed4ae6b3d3e5c3b505