Move flags to base usage library
Summary:
raft_consensus is a monolithic target, and one of the reasons is that it defines a whole bunch of flags. Some of these flags aren't even used in raft_consensus itself. It forces everything to compile with raft_consensus.cc present.

Move these flags to the right place so that the dependency tree is less crazy. For flags used in multi files, follow the standard here: https://gflags.github.io/gflags/#declare

> If you DEFINE a flag in foo.cc, either don't DECLARE it at all, only DECLARE it in tightly related tests, or only DECLARE it in foo.h.

Basically declare the flag in the header so files that import the header can get the flag.

Test Plan: buck build //kudu/...

Reviewed By: li-chi

Differential Revision: D48586961

fbshipit-source-id: bcfe3cb582de513ddd533f2d03c8fff59bd4cc05