Add a debouncer to kuduraft
Summary:
A debouncer is a control structure that let's one party take the a lock, and allows at most one more party to block on the lock. Extra lockers will be rejects.

The idea is that if we have a bunch of repeated calls in succession, the debouncer will "debounce" the duplicates after the second: 1,2,3,4 => 1,2.

This is useful for things like cache updates. Any update triggers a cache refresh. But duplicates while the refresh is ongoing can be collated into a single call after the ongoing refresh completes.

I plan to use this later for filling a buffer for a peer from the log cache while the RPC is ongoing, so that we can read more batches based on the max batch size while waiting for the rpc to complete.

Test Plan: Added a test to make sure it's sane

Reviewed By: alanliang

Differential Revision: D47676502

fbshipit-source-id: e9ec6a9e4d0a4e3ea95ddf4f3e338bed8b903b0f