Fix disable failure detector
Summary:
We bail out the function if flag is off, but some cases we don't set the flag. (Flag was meant to be set at startup).

We should really be able to use disable this idempotently even if it's not started.

Test Plan: `buck test fbcode//mysql_raft/tests:FlexiRaftTest -- FlexiRaftTest.RejectVotesWhenHighLag`

Reviewed By: alanliang

Differential Revision: D48918881

fbshipit-source-id: cb773fbe895b2a42146e0bb70ec556c62aef700b