Create a debounced buffer and add some logic to fill it
Summary:
We create a struct to hold the buffer of messages we're anticipating to send out on the next RPC.

We also add some code to fill this buffer from the last position it was filled to. Filling the buffer is debounced by the debouncer introduced in the previous diff, and enforced with folly::Synchronized, so the debouncer must be locked before we can access the buffer. Debouncing the buffer means only one thread can fill the buffer, and one more can wait. The rest are deduped.

Finally, in this diff we assume the buffer watermark is at some point and we'll just continue filling. We don't have a init code to move it out of -1 at the start, which we will implement in the next diff.

Test Plan: This code path isn't exercised at this point, so compile for now and test together later.

Reviewed By: alanliang

Differential Revision: D47676504

fbshipit-source-id: ccb64320aeba6e8b5df5bdff0ebe379f5af3b388