Implement a mechanism to handoff the buffer to a pending RPC call
Summary:
The idea here is that once we get the signal that we need to send a RPC call, we'll need to extract the buffer that we've been filling up.

To to this, I split the buffer into 2 sections.

The data section is represented by SynchronizedBufferData, and will be synchronized via the deboucing lock.

The handoff section contains an atomic and a promise. When a thread finishes filling the buffer and see that the atomic is set, it'll move the buffered replicates into the promise. The RPC thread will wait for the future to get the buffer. The atomic also serves as the synchronization to ensure indexes are matched. So if the RPC thread provides a different initial index to what we have buffered (happens on the non-happy path for appends), we'll drop the buffer and redo the fill.

The handoff section is already thread safe, so I implemented a custom LockHandle which will lock the data section when alive, while providing APIs into the buffer section for the calling code to use to handoff the buffer.

Test Plan: Nothing is calling this code at this point, we'll test in later diffs.

Reviewed By: alanliang

Differential Revision: D47676503

fbshipit-source-id: a2dc3138d9ae0e6574c96976f8291c9393bd9dcc