Fix race in debouncer-test
Summary:
This test was racy because we wait on the executed latch, then check the number of failed acquires. If the check for the acquire executes before the failed counter is incremented, the test will fail.

Add another latch and make sure we incremented the failed counter.

Test Plan: buck test //kudu/util:debouncer_test -- DebouncerTests.DebouncerTest --run-disabled

Reviewed By: alanliang

Differential Revision: D49034600

fbshipit-source-id: 70c749a89517789e39f7ffdd49394199b8e2f1b3