Initial buffer fill when signaled and get the buffer to send for RPCs
Summary:
Use the functionality implemented in the previous diff to fill the request by getting the handed off buffer.

This is done setting the hand off flags, scheduling a buffer fill, and waiting for the buffer future to get the handed off buffer. Basically, if a fill is ongoing, it'll immediately return the buffer after it completes. If there's no fill ongoing, the one we scheduled will do the fill and handoff.

This diff completes the basic functionality for buffering, in the sense that ops can now be prebuffered if the feature flag is turned on (see test plan).

There are a few more improvements before we can call the feature complete that I'll add in later diffs:

1. Buffer totally ignores max batch size today, so RPC size could become unbounded, especially with a high RTT. We need to stop buffering when we hit max batch size
2. We read from log cache serially for each peer, and attempts are deduped, but if writes are high we could be constantly reading, causing log cache contention. We need some sort of short sleep to batch up the buffers so we can fill more efficiently (or alternatively do something like flat combining on the log cache)
3. If a buffer fill hits the max read limit, it'll just bail out and we'll only fill again on the next transaction. Fill needs to reinvoke fill if it did not manage to fill up all the way.
4.. Proxies have no buffering at all. Which log cache contention is less important on the proxy, if the proxy cannot fill up the whole rpc's worth of ops that we bufferred, we will churn through the buffer as the appended index will never be the same as the index the buffering mechanism is expecting. Proxies should be ok to do multiple smaller reads into log cache to try and fill up the whole RPC's woth of logs.

Test Plan: This is tested manually by adding debug logging and running a Raft ring with high writes. When writes get frequent enough, we should start seeing from debug logs that we can buffer messages in between RPCs and that they will be handed off to the next RPC when a response arrive.

Reviewed By: alanliang

Differential Revision: D47676501

fbshipit-source-id: 67daff533fee9d97e0193a49c2858a143ffc59f1