Track bytes bufferred and continue buffering if we haven't hit max batch
Summary: Up to this point we try to buffer on every appended operation. However, we should also continue bufferring if a single bufferring attempt did not fill up everything from LogCache. We schedule another FillRequest at the end of FillBuffer if we see that we didn't finish reading ops from LogCache, and that we still have bytes left in max batch size.

Test Plan:
Change the logging for refills up to INFO. Lower max_buffer_fill_size_bytes manually.

Upload this to merlin, lower the fill size.

Start a mysqlslap workload and check for the logs showing that we're refilling after the first attempt exits early from ReadOps.

Reviewed By: alanliang

Differential Revision: D48544113

fbshipit-source-id: 8f3b94dd4b44276b1684716058a5008ce608bed9