Disable proxying after error threshold and check periodically to see whether we can re-enable proxying
Summary:
Today, once we are in a proxying mode, the proxy will continue proxying until `now - last_successful_exchange` exceeds some threshold. In the case where a proxy has missing log entries, we degrade the proxy request to a heartbeat. However, this request is still considered a last successful exchange. This means that we will continue proxying with empty messages until the instance gets replaced. It gets replaced because it reports an error to the error manager and ReplicaReplacementTask will react to error and issue a copy replacement.

A better way to handle this would be to disable proxying for some fixed period of time, and send requests directly to the instance. After some period of time, we attempt a proxy request. If it can successfully proxy the request, we continue proxying.

Another benefit of this change is that, this will avoid unnecessary replacements if an instance was recently caught via warm-storage-catchup. In this case, during warm-storage-catchup, the instance is caught up in non-raft mode (raft plugin off), meaning it does not catalog the binlog events that it consumes during catchup. This means that there could be a brief period of time when a newly caught up instance via warm-storage-catchup might not have the necessary logs for a proxy request. This can happen if the proxy target was behind the proxy in terms of appended op ids. With this change, we will allow the proxy target to catch up from leader and then resume from the proxy.

This diff also creates a sysvar to enable reporting of proxy errors to error manager. We eventually want to disable this once we are able to roll out warm-storage-catchups. However, before we can disable this, we still need to resolve the issue where a proxy is so far behind that it is better to issue a copy replacement. This will be addressed in a later diff.

Test Plan:
- Added unit tests.

- Tested in merlindb:
  - simulated a case where target is behind proxy (rejecting appends on proxy target)
    - see P816208762, after this is run, we catchup the lagging mysqld and then remove rejecting appends on the lagging LBU
  - without this diff, proxy will continue proxying forever and degrading to heartbeats until replacement
  - with this diff, proxy is disabled and eventually target catches up and proxying resumes

- `buck2 test @//mode/dev  //mysql_raft/tests:RaftProxyTest -- --stress-runs 200 --jobs 10`

Reviewed By: yichenshen

Differential Revision: D48726861

fbshipit-source-id: 1f25d893ff03c75e254ca7b06189c35091adb70a