Fix infinite loop in log cache
Summary:
Broken in D48453232

There we break out early before decreasing remaining_space. The idea was that it gives an better estimate of how much we read (previously we decrease the value but don't append the operation).

However, because there's a outer while loop that checks remaining_space > 0, we enter an infinite loop.

Revert it first, a couple of tests are breaking. I'll think of how to do proper account later.

Test Plan:
```
buck test //mysql_raft/tests:RaftIndexChunkTest -- --timeout 180
```

Reviewed By: li-chi

Differential Revision: D49115514

fbshipit-source-id: 6b998845c703b9b1b187e2e1173d6f30f9f174ea