Continue reading ops in proxy up till deadline if we didn't read everything we need
Summary:
I was working on the batching piece and I noticed that proxied nodes often need to discard buffers, because they don't append up to the ops that the leader sends it.

We have a deadline to fill the cache, so keep trying to fill up to the leader's request till the deadline. This avoids a bunch of back and forths, and would be helpful for cross altanic proxying too. Currently if we have a mismatch a transaction could take up to 2x cross atlantic RTT to reach the proxied node.

Test Plan:
Use D47676500 to enable batching, then upload together to merlin.

First check that remote LBUs are replicating fine.

Then turn on mysqlslap on the leader to fire some workload and look at the logs.

Previously, we often had logs saying that the buffer does not match the last appended ID for proxies nodes. After this change, these disappear and are replaced with messages that the buffer matches.

Reviewed By: alanliang

Differential Revision: D49474547

fbshipit-source-id: 84a217120dbab05316d3902b72db8e4bc0ec12df